requirements.txt
flask
flask_sqlalchemy
flask_cors
bcrypt
pyjwt
flask-mail
python-dotenv
